refactored winder2's emit methods to emit.rs
